+++ src/App.vue
@@ -1,30 +1,33 @@
 <template>
   <error-boundary>
-    <router-view />
+    <!-- TODO: メニューバーをEditorHomeから移動する -->
+    <router-view v-if="isVuexReady" v-slot="{ Component }">
+      <keep-alive>
+        <component
+          :is="Component"
+          :is-engines-ready="isEnginesReady"
+          :project-file-path="projectFilePath"
+        />
+      </keep-alive>
+    </router-view>
   </error-boundary>
 </template>
 
 <script setup lang="ts">
-import { watch, computed } from "vue";
+import { watch, onMounted, ref, computed } from "vue";
 import { useGtm } from "@gtm-support/vue-gtm";
 import { useRoute } from "vue-router";
+import Mousetrap from "mousetrap";
+import { EngineId } from "./type/preload";
 import ErrorBoundary from "@/components/ErrorBoundary.vue";
 import { useStore } from "@/store";
+import { isOnCommandOrCtrlKeyDown } from "@/store/utility";
 
 const store = useStore();
-store.dispatch("INIT_VUEX");
-
-// URLパラメータに従ってマルチエンジンをオフにする
 const route = useRoute();
-const query = computed(() => route.query);
-watch(query, (newQuery) => {
-  if (newQuery) {
-    store.dispatch(
-      "SET_IS_MULTI_ENGINE_OFF_MODE",
-      newQuery["isMultiEngineOffMode"] === "true"
-    );
-  }
-});
+
+// TODO: プロジェクトファイルの読み込みもEditorHomeから移動する
+const projectFilePath = computed(() => route.query["projectFilePath"]);
 
 // Google Tag Manager
 const gtm = useGtm();
@@ -44,4 +47,80 @@ watch(
   },
   { immediate: true }
 );
+
+// ソフトウェアを初期化
+const isVuexReady = ref(false);
+const isEnginesReady = ref(false);
+onMounted(async () => {
+  const initVuexPromise = store.dispatch("INIT_VUEX");
+
+  // Electronのデフォルトのundo/redoを無効化
+  const disableDefaultUndoRedo = (event: KeyboardEvent) => {
+    // ctrl+z, ctrl+shift+z, ctrl+y
+    if (
+      isOnCommandOrCtrlKeyDown(event) &&
+      (event.key == "z" || (!event.shiftKey && event.key == "y"))
+    ) {
+      event.preventDefault();
+    }
+  };
+  document.addEventListener("keydown", disableDefaultUndoRedo);
+
+  // ショートカットキー操作を止める条件の設定
+  // 止めるなら`true`を返す
+  Mousetrap.prototype.stopCallback = (
+    e: Mousetrap.ExtendedKeyboardEvent, // 未使用
+    element: Element,
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    combo: string // 未使用
+  ) => {
+    return (
+      element.tagName === "INPUT" ||
+      element.tagName === "SELECT" ||
+      element.tagName === "TEXTAREA" ||
+      (element instanceof HTMLElement && element.contentEditable === "true") ||
+      // メニュー項目ではショートカットキーを無効化
+      element.classList.contains("q-item")
+    );
+  };
+
+  await initVuexPromise;
+  isVuexReady.value = true;
+
+  // エンジンの初期化開始
+
+  // エンジン情報取得
+  await store.dispatch("GET_ENGINE_INFOS");
+
+  // URLパラメータに従ってマルチエンジンをオフにする
+  const isMultiEngineOffMode = route.query["isMultiEngineOffMode"] === "true";
+  store.dispatch("SET_IS_MULTI_ENGINE_OFF_MODE", isMultiEngineOffMode);
+
+  // マルチエンジンオフモードのときはデフォルトエンジンだけにする
+  let engineIds: EngineId[];
+  if (isMultiEngineOffMode) {
+    const main = Object.values(store.state.engineInfos).find(
+      (engine) => engine.type === "default"
+    );
+    if (!main) {
+      throw new Error("No main engine found");
+    }
+    engineIds = [main.uuid];
+  } else {
+    engineIds = store.state.engineIds;
+  }
+  await store.dispatch("LOAD_USER_CHARACTER_ORDER");
+  await store.dispatch("POST_ENGINE_START", {
+    engineIds,
+  });
+
+  // 辞書を同期
+  await store.dispatch("SYNC_ALL_USER_DICT");
+
+  isEnginesReady.value = true;
+});
+
+// TODO: ダイアログ周りをEditorHomeから移動する
+
+// TODO: エンジン起動状態周りの処理と表示をEditorHomeから移動する
 </script>
